requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
celery[redis]==5.4.0
redis==5.0.7
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
loguru==0.7.2
tenacity==8.5.0
openai==1.40.0